requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.11.0
arq==0.26.3
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
exceptiongroup==1.2.1
fastapi==0.110.3
fonttools==4.57.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hiredis==3.1.1
idna==3.7
kaleido==0.2.1
kiwisolver==1.4.8
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.0.5
narwhals==1.33.0
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
psycopg2-binary==2.9.9
pyaes==1.6.1
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.2.3
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.3
starlette==0.37.2
TgCrypto==1.2.5
typing_extensions==4.11.0
tzdata==2025.2
tzlocal==5.2
uvicorn==0.29.0
yarl==1.9.4
